Rearrange files.
