+++ languages/rust/Cargo.toml
@@ -1,16 +1,25 @@
 [package]
 name = "oso"
-version = "0.4.1-rc4"
-authors = ["Sam Scott <sam@osohq.com>"]
+version = "0.6.0-rc0"
+authors = ["Oso Security, Inc. <support@osohq.com>"]
 edition = "2018"
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+[[bin]]
+name = "oso"
+path = "src/repl.rs"
+required-features = ["cli"]
 
 [dependencies]
-polar-core = { path = "../../polar-core", version = "0.4.1-rc4" }
-rustyline = {version = "6.2"}
-rustyline-derive = {version = "0.3.1"}
-anyhow = "1.0.32"
 maplit = "1.0.2"
+polar-core = { path = "../../polar-core" }
+thiserror = "1.0.20"
 tracing = "0.1.19"
-tracing-subscriber ={version = "0.2.11", features = ["fmt"] }
+tracing-subscriber = { version = "0.2.11", features = ["fmt"] }
+
+anyhow = { version = "1.0.32", optional = true }
+rustyline = { version = "6.2", optional = true }
+rustyline-derive = { version = "0.3.1", optional = true }
+
+[features]
+default = []
+cli = ["rustyline", "rustyline-derive", "anyhow"]
