+++ languages/rust/src/host/mod.rs
@@ -1,15 +1,20 @@
-//! Translate between Polar and the host language (Rust).
-
-use std::any::Any;
 use std::collections::HashMap;
-use std::rc::Weak;
-use std::sync::Arc;
+use std::rc::Rc;
 
-use polar_core::types::Symbol as Name;
-use polar_core::types::{ExternalInstance, Numeric, Operator, Term, Value};
+use polar_core::terms::{ExternalInstance, Numeric, Operator, Symbol, Term, Value};
 
+use crate::Polar;
+
+mod builtins;
 mod class;
-pub use class::*;
+mod class_method;
+mod from_polar;
+mod method;
+mod to_polar;
+
+pub use class::{Class, Instance};
+pub use from_polar::FromPolar;
+pub use to_polar::ToPolar;
 
 #[derive(Clone, Default)]
 pub struct Type;
@@ -19,16 +24,16 @@ pub fn type_class() -> Class {
     class.erase_type()
 }
 
-/// Maintain mappings and caches for Python classes & instances
+/// Maintain mappings and caches for Rust classes & instances
 pub struct Host {
-    polar: Weak<crate::PolarCore>,
-    classes: HashMap<Name, Class>,
-    instances: HashMap<u64, Instance>,
-    class_names: HashMap<std::any::TypeId, Name>,
+    polar: Rc<Polar>,
+    classes: HashMap<Symbol, Class>,
+    instances: HashMap<u64, class::Instance>,
+    class_names: HashMap<std::any::TypeId, Symbol>,
 }
 
 impl Host {
-    pub fn new(polar: Weak<crate::PolarCore>) -> Self {
+    pub fn new(polar: Rc<Polar>) -> Self {
         let mut host = Self {
             class_names: HashMap::new(),
             classes: HashMap::new(),
@@ -36,17 +41,24 @@ impl Host {
             polar,
         };
         let type_class = type_class();
-        let name = Name("Type".to_string());
+        let name = Symbol("Type".to_string());
         host.class_names.insert(type_class.type_id, name.clone());
         host.classes.insert(name, type_class);
+
+        // register all builtin constants
+        for (name, class) in builtins::classes() {
+            let value = host.cache_class(class, name.clone());
+            host.polar.register_constant(name, value);
+        }
+
         host
     }
 
     pub fn type_class(&mut self) -> &mut Class {
-        self.classes.get_mut(&Name("Type".to_string())).unwrap()
+        self.classes.get_mut(&Symbol("Type".to_string())).unwrap()
     }
 
-    pub fn get_class(&self, name: &Name) -> Option<&Class> {
+    pub fn get_class(&self, name: &Symbol) -> Option<&Class> {
         self.classes.get(name)
     }
 
@@ -56,14 +68,14 @@ impl Host {
             .and_then(|name| self.get_class(name))
     }
 
-    pub fn get_class_mut(&mut self, name: &Name) -> Option<&mut Class> {
+    pub fn get_class_mut(&mut self, name: &Symbol) -> Option<&mut Class> {
         self.classes.get_mut(name)
     }
 
     /// Add the class to the host classes
     ///
     /// Returns an instance of `Type` for this class.
-    pub fn cache_class(&mut self, class: Class, name: Name) -> Term {
+    pub fn cache_class(&mut self, class: Class, name: Symbol) -> Term {
         self.class_names.insert(class.type_id, name.clone());
         self.classes.insert(name, class.clone());
 
@@ -73,7 +85,7 @@ impl Host {
                 .instance_methods
                 .entry(method_name.clone())
                 .or_insert_with(|| {
-                    crate::host::InstanceMethod::from_class_method(method_name.clone())
+                    class_method::InstanceMethod::from_class_method(method_name.clone())
                 });
         }
         let repr = format!("type<{}>", class.name);
@@ -86,17 +98,17 @@ impl Host {
         }))
     }
 
-    pub fn get_instance(&self, id: u64) -> Option<&Instance> {
+    pub fn get_instance(&self, id: u64) -> Option<&class::Instance> {
         self.instances.get(&id)
     }
 
-    pub fn cache_instance(&mut self, instance: Instance, id: Option<u64>) -> u64 {
-        let id = id.unwrap_or_else(|| self.polar.upgrade().unwrap().get_external_id());
+    pub fn cache_instance(&mut self, instance: class::Instance, id: Option<u64>) -> u64 {
+        let id = id.unwrap_or_else(|| self.polar.get_external_id());
         self.instances.insert(id, instance);
         id
     }
 
-    pub fn make_instance(&mut self, name: &Name, fields: Vec<Term>, id: u64) {
+    pub fn make_instance(&mut self, name: &Symbol, fields: Vec<Term>, id: u64) {
         let class = self.get_class(name).unwrap().clone();
         debug_assert!(self.instances.get(&id).is_none());
         let fields = fields; // TODO: use
@@ -110,7 +122,7 @@ impl Host {
         todo!("left == right")
     }
 
-    pub fn isa(&self, term: Term, class_tag: &Name) -> bool {
+    pub fn isa(&self, term: Term, class_tag: &Symbol) -> bool {
         let name = &class_tag.0;
         match term.value() {
             Value::ExternalInstance(ExternalInstance { instance_id, .. }) => {
@@ -133,7 +145,7 @@ impl Host {
         }
     }
 
-    pub fn is_subspecializer(&self, id: u64, left_tag: &Name, right_tag: &Name) -> bool {
+    pub fn is_subspecializer(&self, id: u64, left_tag: &Symbol, right_tag: &Symbol) -> bool {
         let _instance = self.get_instance(id).unwrap();
         let _left = self.get_class(left_tag).unwrap();
         let _right = self.get_class(right_tag).unwrap();
@@ -141,7 +153,7 @@ impl Host {
         todo!("????")
     }
 
-    pub fn operator(&self, _op: Operator, _args: [Instance; 2]) -> bool {
+    pub fn operator(&self, _op: Operator, _args: [class::Instance; 2]) -> bool {
         todo!()
     }
 
@@ -150,327 +162,5 @@ impl Host {
     }
 }
 
-pub trait ToPolar {
-    fn to_polar_value(&self, host: &mut Host) -> Value;
-
-    fn to_polar(&self, host: &mut Host) -> Term {
-        Term::new_from_ffi(self.to_polar_value(host))
-    }
-}
-
-impl ToPolar for bool {
-    fn to_polar_value(&self, _host: &mut Host) -> Value {
-        Value::Boolean(*self)
-    }
-}
-
-macro_rules! int_to_polar {
-    ($i:ty) => {
-        impl ToPolar for $i {
-            fn to_polar_value(&self, _host: &mut Host) -> Value {
-                Value::Number(Numeric::Integer((*self).into()))
-            }
-        }
-    };
-}
-
-int_to_polar!(u8);
-int_to_polar!(i8);
-int_to_polar!(u16);
-int_to_polar!(i16);
-int_to_polar!(u32);
-int_to_polar!(i32);
-int_to_polar!(i64);
-
-macro_rules! float_to_polar {
-    ($i:ty) => {
-        impl ToPolar for $i {
-            fn to_polar_value(&self, _host: &mut Host) -> Value {
-                Value::Number(Numeric::Float((*self).into()))
-            }
-        }
-    };
-}
-
-float_to_polar!(f32);
-float_to_polar!(f64);
-
-impl ToPolar for String {
-    fn to_polar_value(&self, _host: &mut Host) -> Value {
-        Value::String(self.clone())
-    }
-}
-
-impl ToPolar for &'static str {
-    fn to_polar_value(&self, _host: &mut Host) -> Value {
-        Value::String(self.to_string())
-    }
-}
-
-impl ToPolar for str {
-    fn to_polar_value(&self, _host: &mut Host) -> Value {
-        Value::String(self.to_owned())
-    }
-}
-
-impl<T: ToPolar> ToPolar for Vec<T> {
-    fn to_polar_value(&self, host: &mut Host) -> Value {
-        Value::List(self.iter().map(|v| v.to_polar(host)).collect())
-    }
-}
-
-impl<T: ToPolar> ToPolar for HashMap<String, T> {
-    fn to_polar_value(&self, host: &mut Host) -> Value {
-        Value::Dictionary(polar_core::types::Dictionary {
-            fields: self
-                .iter()
-                .map(|(k, v)| (Name(k.to_string()), v.to_polar(host)))
-                .collect(),
-        })
-    }
-}
-
-pub struct PolarIter<I>(pub I);
-
-impl<I: Clone + Iterator<Item = T>, T: ToPolar> ToPolar for PolarIter<I> {
-    fn to_polar_value(&self, host: &mut Host) -> Value {
-        Value::List(self.0.clone().map(|v| v.to_polar(host)).collect())
-    }
-}
-
-impl ToPolar for Value {
-    fn to_polar_value(&self, _host: &mut Host) -> Value {
-        self.clone()
-    }
-}
-
-impl ToPolar for Box<dyn ToPolar> {
-    fn to_polar_value(&self, host: &mut Host) -> Value {
-        self.as_ref().to_polar_value(host)
-    }
-}
-
-pub trait FromPolar: Sized {
-    fn from_polar(term: &Term, host: &mut Host) -> Option<Self>;
-
-    fn from_polar_list(terms: &[Term], host: &mut Host) -> Option<Self> {
-        assert_eq!(terms.len(), 1);
-        Self::from_polar(&terms[0], host)
-    }
-}
-
-impl FromPolar for bool {
-    fn from_polar(term: &Term, _host: &mut Host) -> Option<Self> {
-        if let Value::Boolean(b) = term.value() {
-            Some(*b)
-        } else {
-            None
-        }
-    }
-}
-
-use std::convert::TryFrom;
-
-macro_rules! polar_to_int {
-    ($i:ty) => {
-        impl FromPolar for $i {
-            fn from_polar(term: &Term, _host: &mut Host) -> Option<Self> {
-                if let Value::Number(Numeric::Integer(i)) = term.value() {
-                    <$i>::try_from(*i).ok()
-                } else {
-                    None
-                }
-            }
-        }
-    };
-}
-
-polar_to_int!(u8);
-polar_to_int!(i8);
-polar_to_int!(u16);
-polar_to_int!(i16);
-polar_to_int!(u32);
-polar_to_int!(i32);
-polar_to_int!(i64);
-
-impl FromPolar for f64 {
-    fn from_polar(term: &Term, _host: &mut Host) -> Option<Self> {
-        if let Value::Number(Numeric::Float(f)) = term.value() {
-            Some(*f)
-        } else {
-            None
-        }
-    }
-}
-
-impl FromPolar for String {
-    fn from_polar(term: &Term, _host: &mut Host) -> Option<Self> {
-        if let Value::String(s) = term.value() {
-            Some(s.to_string())
-        } else {
-            None
-        }
-    }
-}
-
-impl<T: FromPolar> FromPolar for Vec<T> {
-    fn from_polar(term: &Term, host: &mut Host) -> Option<Self> {
-        if let Value::List(l) = term.value() {
-            l.iter().map(|t| T::from_polar(t, host)).collect()
-        } else {
-            None
-        }
-    }
-
-    fn from_polar_list(terms: &[Term], host: &mut Host) -> Option<Self> {
-        terms.iter().map(|t| T::from_polar(t, host)).collect()
-    }
-}
-
-impl<T: FromPolar> FromPolar for HashMap<String, T> {
-    fn from_polar(term: &Term, host: &mut Host) -> Option<Self> {
-        if let Value::Dictionary(dict) = term.value() {
-            dict.fields
-                .iter()
-                .map(|(k, v)| T::from_polar(v, host).map(|v| (k.0.clone(), v)))
-                .collect()
-        } else {
-            None
-        }
-    }
-}
-
-impl FromPolar for Value {
-    fn from_polar(term: &Term, _host: &mut Host) -> Option<Self> {
-        Some(term.value().clone())
-    }
-}
-
-impl FromPolar for Instance {
-    fn from_polar(term: &Term, host: &mut Host) -> Option<Self> {
-        let instance = match term.value().clone() {
-            Value::Boolean(b) => host
-                .get_class_from_type::<bool>()
-                .unwrap()
-                .cast_to_instance(b),
-            Value::Number(Numeric::Integer(i)) => host
-                .get_class_from_type::<i64>()
-                .unwrap()
-                .cast_to_instance(i),
-            Value::Number(Numeric::Float(f)) => host
-                .get_class_from_type::<f64>()
-                .unwrap()
-                .cast_to_instance(f),
-            Value::List(v) => host
-                .get_class_from_type::<Vec<Term>>()
-                .unwrap()
-                .cast_to_instance(v),
-            Value::String(s) => host
-                .get_class_from_type::<String>()
-                .unwrap()
-                .cast_to_instance(s),
-            Value::Dictionary(d) => host
-                .get_class_from_type::<HashMap<Name, Term>>()
-                .unwrap()
-                .cast_to_instance(d.fields),
-            Value::ExternalInstance(ExternalInstance { instance_id, .. }) => host
-                .get_instance(instance_id)
-                .expect("instance not found")
-                .clone(),
-            v => {
-                tracing::warn!(value = ?v, "invalid conversion attempted");
-                return None;
-            }
-        };
-        Some(instance)
-    }
-}
-
-impl FromPolar for () {
-    fn from_polar(term: &Term, _host: &mut Host) -> Option<Self> {
-        if let Value::List(l) = term.value() {
-            if l.is_empty() {
-                Some(())
-            } else {
-                None
-            }
-        } else {
-            None
-        }
-    }
-
-    fn from_polar_list(terms: &[Term], _host: &mut Host) -> Option<Self> {
-        if terms.is_empty() {
-            Some(())
-        } else {
-            None
-        }
-    }
-}
-
-impl<A> FromPolar for (A,)
-where
-    A: FromPolar,
-{
-    fn from_polar(_term: &Term, _host: &mut Host) -> Option<Self> {
-        None
-    }
-
-    fn from_polar_list(terms: &[Term], host: &mut Host) -> Option<Self> {
-        if terms.len() == 1 {
-            A::from_polar(&terms[0], host).map(|a| (a,))
-        } else {
-            None
-        }
-    }
-}
-
-impl<A, B> FromPolar for (A, B)
-where
-    A: FromPolar,
-    B: FromPolar,
-{
-    fn from_polar(_term: &Term, _host: &mut Host) -> Option<Self> {
-        None
-    }
-
-    fn from_polar_list(terms: &[Term], host: &mut Host) -> Option<Self> {
-        if terms.len() == 2 {
-            let a = A::from_polar(&terms[0], host)?;
-            let b = B::from_polar(&terms[1], host)?;
-            Some((a, b))
-        } else {
-            None
-        }
-    }
-}
-
 /// Marker trait: implements "ToPolar" via a registered class
 pub trait HostClass {}
-
-impl<C: 'static + Clone + HostClass> FromPolar for C {
-    fn from_polar(term: &Term, host: &mut Host) -> Option<Self> {
-        match term.value() {
-            Value::ExternalInstance(ExternalInstance { instance_id, .. }) => {
-                let instance = host.get_instance(*instance_id)?;
-                instance.instance.downcast_ref::<C>().cloned()
-            }
-            _ => None,
-        }
-    }
-}
-
-impl<C: 'static + Clone + HostClass> ToPolar for C {
-    fn to_polar_value(&self, host: &mut Host) -> Value {
-        let class = host
-            .get_class_from_type::<C>()
-            .expect("Class not registered");
-        let instance = class.cast_to_instance(self.clone());
-        let instance = host.cache_instance(instance, None);
-        Value::ExternalInstance(ExternalInstance {
-            constructor: None,
-            repr: None,
-            instance_id: instance,
-        })
-    }
-}
