+++ languages/rust/src/oso.rs
@@ -1,0 +1,138 @@
+//! Communicate with the Polar virtual machine: load rules, make queries, etc/
+
+use polar_core::terms::{Predicate, Symbol, Term, Value};
+
+use std::fs::File;
+use std::io::Read;
+use std::rc::Rc;
+use std::sync::{Arc, Mutex};
+
+use crate::host::Host;
+use crate::query::Query;
+use crate::ToPolar;
+
+#[derive(Clone)]
+pub struct Oso {
+    inner: Rc<polar_core::polar::Polar>,
+    host: Arc<Mutex<Host>>,
+}
+
+impl Default for Oso {
+    fn default() -> Self {
+        Self::new()
+    }
+}
+
+impl Oso {
+    pub fn new() -> Self {
+        let inner = Rc::new(polar_core::polar::Polar::new());
+        let host = Host::new(inner.clone());
+
+        Self {
+            host: Arc::new(Mutex::new(host)),
+            inner,
+        }
+    }
+
+    pub fn is_allowed<Actor, Action, Resource>(
+        &mut self,
+        actor: Actor,
+        action: Action,
+        resource: Resource,
+    ) -> bool
+    where
+        Actor: ToPolar,
+        Action: ToPolar,
+        Resource: ToPolar,
+    {
+        let args: Vec<&dyn ToPolar> = vec![&actor, &action, &resource];
+        let mut query = self.query_rule("allow", args).unwrap();
+        query.next().is_some()
+    }
+
+    pub fn clear(&mut self) {
+        *self = Self::new();
+    }
+
+    fn check_messages(&mut self) {
+        while let Some(msg) = self.inner.next_message() {
+            match msg.kind {
+                polar_core::messages::MessageKind::Print => tracing::trace!("{}", msg.msg),
+                polar_core::messages::MessageKind::Warning => tracing::warn!("{}", msg.msg),
+            }
+        }
+    }
+
+    fn check_inline_queries(&mut self) -> crate::Result<()> {
+        while let Some(q) = self.inner.next_inline_query(false) {
+            let query = Query::new(q, self.host.clone());
+            match query.collect::<crate::Result<Vec<_>>>() {
+                Ok(v) if !v.is_empty() => continue,
+                Ok(_) => return lazy_error!("inline query result was false"),
+                Err(e) => return lazy_error!("error in inline query: {}", e),
+            }
+        }
+        self.check_messages();
+        Ok(())
+    }
+
+    pub fn load_file(&mut self, file: &str) -> crate::Result<()> {
+        let mut f = File::open(&file)?;
+        let mut policy = String::new();
+        f.read_to_string(&mut policy)?;
+        self.inner.load(&policy, Some(file.to_string()))?;
+        self.check_inline_queries()
+    }
+
+    pub fn load_str(&mut self, s: &str) -> crate::Result<()> {
+        self.inner.load(s, None)?;
+        self.check_inline_queries()
+    }
+
+    pub fn query(&mut self, s: &str) -> crate::Result<Query> {
+        let query = self.inner.new_query(s, false)?;
+        self.check_messages();
+        let query = Query::new(query, self.host.clone());
+        Ok(query)
+    }
+
+    pub fn query_rule<'a>(
+        &mut self,
+        name: &str,
+        args: impl IntoIterator<Item = &'a dyn crate::host::ToPolar>,
+    ) -> crate::Result<Query> {
+        let args = args
+            .into_iter()
+            .map(|arg| arg.to_polar(&mut self.host.lock().unwrap()))
+            .collect();
+        let query_value = Value::Call(Predicate {
+            name: Symbol(name.to_string()),
+            args,
+        });
+        let query_term = Term::new_from_ffi(query_value);
+        let query = self.inner.new_query_from_term(query_term, false);
+        self.check_messages();
+        let query = Query::new(query, self.host.clone());
+        Ok(query)
+    }
+
+    pub fn register_class(&mut self, class: crate::host::Class) -> crate::Result<()> {
+        let mut host = self.host.lock().unwrap();
+        let name = class.name.clone();
+        let name = Symbol(name);
+        let class_term = host.cache_class(class, name.clone());
+        self.inner.register_constant(name, class_term);
+        Ok(())
+    }
+
+    pub fn register_constant<V: crate::host::ToPolar>(
+        &mut self,
+        name: &str,
+        value: &V,
+    ) -> crate::Result<()> {
+        let mut host = self.host.lock().unwrap();
+        self.inner
+            .register_constant(Symbol(name.to_string()), host.value_to_polar(value));
+        Ok(())
+    }
+}
